requirements.txt
discord.py
dotenv
yt_dlp